SUPERNOVA: Adds resource abstraction
The resource abstraction includes loading of sound files and cursor
graphics.
